add session page

+++ ex01/Cinema/src/main/java/com/admiralxy/cinema/repositories/interfaces/ISessionsRepository.java
@@ -4,7 +4,7 @@ import com.admiralxy.cinema.entities.Session;
 import java.util.List;
 
 public interface ISessionsRepository {
-    Session findById(int id);
+    Session findById(Long id);
     List<Session> findAll();
     List<Session> findByFilmTitle(String title);
     void save(Session session);
