+++ ex01/Cinema/src/main/java/com/admiralxy/cinema/services/interfaces/ISessionsService.java
@@ -6,7 +6,7 @@ import com.admiralxy.cinema.entities.Session;
 import java.util.List;
 
 public interface ISessionsService {
-    Session findById(int id);
+    Session findById(Long id);
     List<Session> findAll();
     List<Session> findByFilmTitle(String title);
     void save(SessionCreateDTO session);
